requirements.txt
Package           Version
----------------- -------
aniso8601         9.0.1
click             7.1.2
Flask             1.1.2
flask-marshmallow 0.14.0
Flask-RESTful     0.3.8
itsdangerous      1.1.0
Jinja2            2.11.3
MarkupSafe        1.1.1
marshmallow       3.11.1
pip               21.1.1
pymongo           3.11.3
python-dotenv     0.17.1
pytz              2021.1
setuptools        56.0.0
six               1.15.0
Werkzeug          1.0.1
